feat: 完成 Note CRUD API ResponseEntity 和 HTTP 狀態碼改寫

+++ src/main/java/com/jeannychiu/learningnotesapi/service/NoteService.java
@@ -46,7 +46,12 @@ public class NoteService {
         return null;
     }
 
-    public void deleteNote(Long id) {
-        noteRepository.deleteById(id);
+    public boolean deleteNote(Long id) {
+        if (noteRepository.existsById(id)) {
+            noteRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
